chore(mini-todo): Change package structure (#43)
