+++ src/main/java/eezn/todolist/minitodo/AppConfig.java
@@ -1,13 +1,13 @@
 package eezn.todolist.minitodo;
 
-import eezn.todolist.minitodo.repository.CategoryRepository;
-import eezn.todolist.minitodo.repository.PriorityRepository;
-import eezn.todolist.minitodo.repository.TodoRepository;
-import eezn.todolist.minitodo.repository.UserRepository;
-import eezn.todolist.minitodo.repository.jdbctemplate.JdbcTemplateCategoryRepository;
-import eezn.todolist.minitodo.repository.jdbctemplate.JdbcTemplatePriorityRepository;
-import eezn.todolist.minitodo.repository.jdbctemplate.JdbcTemplateTodoRepository;
-import eezn.todolist.minitodo.repository.jdbctemplate.JdbcTemplateUserRepository;
+import eezn.todolist.minitodo.domain.todolist.options.repository.CategoryRepository;
+import eezn.todolist.minitodo.domain.todolist.options.repository.PriorityRepository;
+import eezn.todolist.minitodo.domain.todolist.todo.repository.TodoRepository;
+import eezn.todolist.minitodo.domain.account.user.repository.UserRepository;
+import eezn.todolist.minitodo.domain.todolist.options.repository.JdbcTemplateCategoryRepository;
+import eezn.todolist.minitodo.domain.todolist.options.repository.JdbcTemplatePriorityRepository;
+import eezn.todolist.minitodo.domain.todolist.todo.repository.JdbcTemplateTodoRepository;
+import eezn.todolist.minitodo.domain.account.user.repository.JdbcTemplateUserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
