--- src/main/java/eezn/todolist/minitodo/JdbcTemplateConfig.java
@@ -1,17 +1,0 @@
-package eezn.todolist.minitodo;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.jdbc.core.JdbcTemplate;
-
-import javax.sql.DataSource;
-
-@Configuration
-public class JdbcTemplateConfig {
-
-    @Bean
-    public JdbcTemplate jdbcTemplate(DataSource dataSource) {
-        return new JdbcTemplate(dataSource);
-    }
-}
-
