--- src/main/java/eezn/todolist/minitodo/common/constants/TodoConstant.java
@@ -1,16 +1,0 @@
-package eezn.todolist.minitodo.common.constants;
-
-public class TodoConstant {
-
-    public static final int CATEGORY_DEFAULT = 1;
-    public static final int CATEGORY_WORK = 2;
-    public static final int CATEGORY_LIFE = 3;
-
-    public static final int PRIORITY_A = 1;
-    public static final int PRIORITY_B = 2;
-    public static final int PRIORITY_C = 3;
-    public static final int PRIORITY_D = 4;
-
-    public static final int STATUS_TODO = 1;
-    public static final int STATUS_DONE = 2;
-}
