--- src/main/java/eezn/todolist/minitodo/controller/dto/UserDto.java
@@ -1,12 +1,0 @@
-package eezn.todolist.minitodo.controller.dto;
-
-import lombok.Data;
-
-@Data
-public class UserDto {
-
-    private Integer userId;
-    private String username;
-    private String password;
-    private String email;
-}
