+++ src/main/java/eezn/todolist/minitodo/domain/todolist/options/data/Status.java
@@ -1,0 +1,22 @@
+package eezn.todolist.minitodo.domain.todolist.options.data;
+
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
+
+/**
+ *  `status_id`	        INT	            NOT NULL    AUTO_INCREMENT,
+ * 	`status_name`	    VARCHAR(255)	NULL,
+ * 	`status_is_deleted`	BOOLEAN	        NULL    DEFAULT FALSE
+ */
+
+@Getter
+@Setter
+@ToString
+public class Status {
+
+    private Integer id;
+
+    private String status;
+    private Boolean isDeleted;
+}
