--- src/main/java/eezn/todolist/minitodo/repository/CategoryRepository.java
@@ -1,15 +1,0 @@
-package eezn.todolist.minitodo.repository;
-
-import eezn.todolist.minitodo.domain.Category;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface CategoryRepository {
-
-    Category insert(Category category);
-
-    Optional<Category> findById(Integer id);
-
-    List<Category> findAll();
-}
