+++ src/main/resources/application.properties
@@ -5,7 +5,7 @@ spring.datasource.embedded-database-connection=h2
 spring.h2.console.enabled=true
 spring.h2.console.path=/h2-console
 spring.sql.init.schema-locations=classpath*:db/schema.sql
-spring.sql.init.data-locations=classpath*:db/data.sql
+spring.sql.init.data-locations=classpath*:db/test_dummy.sql
 
 # MariaDB(Docker)
 #spring.datasource.driverClassName=org.mariadb.jdbc.Driver
