+++ src/test/java/eezn/todolist/minitodo/repository/JdbcTemplateUserFindTest.java
@@ -1,16 +1,5 @@
 package eezn.todolist.minitodo.repository;
 
-import eezn.todolist.minitodo.AppConfig;
-import eezn.todolist.minitodo.domain.User;
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Test;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseBuilder;
-import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseType;
-
-import javax.sql.DataSource;
-
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class JdbcTemplateUserFindTest {
