+++ src/test/java/eezn/todolist/minitodo/service/UserServiceTest.java
@@ -1,16 +1,5 @@
 package eezn.todolist.minitodo.service;
 
-import eezn.todolist.minitodo.AppConfig;
-import eezn.todolist.minitodo.domain.User;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Test;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseBuilder;
-import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseType;
-
-import javax.sql.DataSource;
-
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
